Fix sound device: use u32 GetChunk override to match Circle
Circle's PWM and I2S sound devices use u32* buffers, not s16*.
Our GetChunk(s16*) override wasn't being called - the base class
GetChunk(u32*) was used instead, which requires AllocateQueue().

This caused 'assertion failed: m_nQueueSize > 1' at startup.

Fix:
- Override GetChunk(u32*, ...) for both CPdSoundPWM and CPdSoundI2S
- Convert libpd float output to device-specific range using
  GetRangeMin()/GetRangeMax() for proper PWM/I2S levels

+++ src/pdsounddevice.h
@@ -47,7 +47,8 @@ public:
 	unsigned GetOutputChannels (void) const { return m_nOutChannels; }
 
 protected:
-	unsigned GetChunk (s16 *pBuffer, unsigned nChunkSize) override;
+	// PWM uses 32-bit samples (u32), not 16-bit
+	unsigned GetChunk (u32 *pBuffer, unsigned nChunkSize) override;
 
 private:
 	float *m_pInBuffer;
@@ -79,7 +80,8 @@ public:
 	unsigned GetOutputChannels (void) const { return m_nOutChannels; }
 
 protected:
-	unsigned GetChunk (s16 *pBuffer, unsigned nChunkSize) override;
+	// I2S uses 32-bit samples (u32)
+	unsigned GetChunk (u32 *pBuffer, unsigned nChunkSize) override;
 
 private:
 	float *m_pInBuffer;
